Update spm_reader.hpp
Added the function of saving the byte data in the spm reader object to the spm file. The current application of this function is mainly used for image channel separation of spm files, which can be achieved by first loading the image data of one or several channels and then saving them.